fixed issues with gtest, catch2, & conan
If you do not indicate to find_package to first look in the bin 
directory, then conan might not be found first when looking up gtest or 
catch2.

Also cleaned up the find_package, as not indicating module or config 
means it will look for both.

Conan has a fix, in case the user file conanfile.txt does not indicate a 
cmake generator.