Create yacht.sh
Combines all the previous scripts, and the bash script in tools as a single self contained script.

This should make things easier to use, and a single entry point to work with the code.

The idea is to have something that will eventually be as close as possible as `cargo` in Rust.
But simpler and not as featured.